Renamed fastoad.utils into fastoad._utils
