+++ src/fastoad/_utils/__init__.py
@@ -1,9 +1,9 @@
+# -*- coding: utf-8 -*-
 """
-Convenience functions for file and directories
+This package contains various utilities
 """
-
 #  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
-#  Copyright (C) 2020  ONERA & ISAE-SUPAERO
+#  Copyright (C) 2021 ONERA & ISAE-SUPAERO
 #  FAST is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
 #  the Free Software Foundation, either version 3 of the License, or
@@ -14,13 +14,3 @@ Convenience functions for file and directories
 #  GNU General Public License for more details.
 #  You should have received a copy of the GNU General Public License
 #  along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-import os
-import os.path as pth
-
-
-def make_parent_dir(path: str):
-    """Ensures parent directory of provided path exists or is created"""
-    dirname = pth.abspath(pth.dirname(path))
-    if not pth.exists(dirname):
-        os.makedirs(dirname)
