+++ src/fastoad/_utils/resource_management/__init__.py
@@ -1,9 +1,5 @@
-# -*- coding: utf-8 -*-
-"""
-This package contains various utilities
-"""
 #  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
-#  Copyright (C) 2020  ONERA & ISAE-SUPAERO
+#  Copyright (C) 2021 ONERA & ISAE-SUPAERO
 #  FAST is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
 #  the Free Software Foundation, either version 3 of the License, or
