+++ src/fastoad/_utils/resource_management/tests/resources/python-logo@2x.png
