+++ src/fastoad/_utils/resource_management/tests/resources/subfolder/py.png
