+++ src/fastoad/io/xml/variable_io_base.py
@@ -2,7 +2,7 @@
 Defines how OpenMDAO variables are serialized to XML using a conversion table
 """
 #  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
-#  Copyright (C) 2020  ONERA & ISAE-SUPAERO
+#  Copyright (C) 2021 ONERA & ISAE-SUPAERO
 #  FAST is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
 #  the Free Software Foundation, either version 3 of the License, or
@@ -22,22 +22,24 @@ from typing import IO, Union
 
 import numpy as np
 from lxml import etree
-from lxml.etree import XPathEvalError
-from lxml.etree import _Element  # pylint: disable=protected-access  # Useful for type hinting
+from lxml.etree import (
+    XPathEvalError,
+    _Element,
+)  # pylint: disable=protected-access  # Useful for type hinting
 from openmdao.vectors.vector import Vector
 
+from fastoad._utils.files import make_parent_dir
+from fastoad._utils.strings import get_float_list_from_string
 from fastoad.io.formatter import IVariableIOFormatter
 from fastoad.io.xml.exceptions import (
     FastXPathEvalError,
-    FastXpathTranslatorXPathError,
-    FastXpathTranslatorVariableError,
     FastXmlFormatterDuplicateVariableError,
+    FastXpathTranslatorVariableError,
+    FastXpathTranslatorXPathError,
 )
 from fastoad.io.xml.translator import VarXpathTranslator
 from fastoad.openmdao.variables import VariableList
-from fastoad.utils.files import make_parent_dir
-from fastoad.utils.strings import get_float_list_from_string
-from .constants import DEFAULT_UNIT_ATTRIBUTE, DEFAULT_IO_ATTRIBUTE, ROOT_TAG
+from .constants import DEFAULT_IO_ATTRIBUTE, DEFAULT_UNIT_ATTRIBUTE, ROOT_TAG
 
 _LOGGER = logging.getLogger(__name__)  # Logger for this module
