+++ src/fastoad/models/aerodynamics/external/xfoil/xfoil_polar.py
@@ -2,7 +2,7 @@
 This module launches XFOIL computations
 """
 #  This file is part of FAST-OAD : A framework for rapid Overall Aircraft Design
-#  Copyright (C) 2020  ONERA & ISAE-SUPAERO
+#  Copyright (C) 2021 ONERA & ISAE-SUPAERO
 #  FAST is free software: you can redistribute it and/or modify
 #  it under the terms of the GNU General Public License as published by
 #  the Free Software Foundation, either version 3 of the License, or
@@ -26,9 +26,9 @@ import numpy as np
 from openmdao.components.external_code_comp import ExternalCodeComp
 from openmdao.utils.file_wrap import InputFileGenerator
 
+from fastoad._utils.resource_management.copy import copy_resource
 from fastoad.models.aerodynamics.external.xfoil import xfoil699
 from fastoad.models.geometry.profiles import get_profile
-from fastoad.utils.resource_management.copy import copy_resource
 from . import resources
 
 OPTION_RESULT_POLAR_FILENAME = "result_polar_filename"
